Less code for mapFloorTree
