+++ include/mapFloorTreeWidget.hpp
@@ -22,22 +22,6 @@ namespace MapFloorsList {
 class MapFloorTreeModel;
 class MapTreeItem;
 
-
-// This line shouldn't even exist. Still here is the reason why:
-//
-// On the signals/slots Qt mechanism, the connect() mechanism is based on
-// textual comparison.
-// Thus, the slot cannot be "void toggleLayerVisibleState(::QModelIndex);"
-// since Qt Designer does not seem to like it.
-// Instead, one has to set "void toggleLayerVisibleState(QModelIndex);".
-// Though, "QModelIndex" would be considered here to be in the
-// "MapFloorsList" namespace (if it ever exists). This is ugly.
-// So the following directive will create an alias that points on the
-// global namespace's QModelIndex.
-//
-// That's it.
-using QModelIndex = ::QModelIndex;
-
 //////////////////////////////////////////////////////////////////////////////
 //  Widget class
 // This widget display a tree-view of the list of all floors of a map
@@ -49,23 +33,16 @@ class FloorListWidget : public QWidget
 public:
     explicit FloorListWidget(QWidget* parent = nullptr);
 
-    const MapFloorTreeModel* mapFloorTreeModel() const;
-
     void setEditorMap(std::shared_ptr<Editor::Map>);
-
-public slots:
-    void selectLayer(QModelIndex);
-    void toggleLayerVisibleState(QModelIndex);
-    void on_treeViewFloors_clicked(const QModelIndex& index);
-
-signals:
-    void treeItemSelected(const MapTreeItem&);
-
-private:
     void reset();
 
-    Ui::MapFloorsList* m_ui             = nullptr;
-    MapFloorTreeModel* m_floorTreeModel = nullptr;
+public slots:
+    void on_treeViewFloors_clicked(const QModelIndex& index);
+    void on_treeViewFloors_doubleClicked(const QModelIndex& index);
+
+private:
+    std::unique_ptr<Ui::MapFloorsList> m_ui;
+    std::unique_ptr<MapFloorTreeModel> m_floorTreeModel;
 };
 
 } // namespace MapFloorsList
